requirements.txt
fastapi
uvicorn
bcrypt
jinja2
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
transformers
torch
certifi